Namespace nested exports resource
The goal is to avoid devs thinking that this is a generic controller
that deals with all kinds of exports.
